Adds support to writing to the database

+++ classes/Image.inc.php
@@ -1,16 +1,16 @@
 <?php
-
-
+include('db.inc.php');
 
 class Image {
 	private $file;
-	private $tmpfile;
-	private $filemime;
+	private $tmpFile;
+	private $fileMIME;
 	private $fileExtension;
+	private $fileSize;
 
 	private $filedescription;
 
-	private $filename;
+	private $originalname;
 
 	// Eventually users should be able to mark something as a duplicate
 	// and then the files can be merged. Filehashes need to be able to be
@@ -23,10 +23,9 @@ class Image {
 	}
 
 	private function processImage() {
-		$this->filename = $this->file['name'];
+		$this->originalname = $this->file['name'];
 		$this->tmpFile = $this->file['tmp_name'];
-		$this->filemime = $this->file['type'];
-		if ($this->checkFileMIME()) {
+		if ($this->checkFileMIME($this->file['type'])) {
 			$this->saveFile();
 		} else {
 			die('Illegal file type');
@@ -36,38 +35,49 @@ class Image {
 	private function saveFile() {
 		$hash = sha1_file($this->tmpFile);
 		$destination = ROOTDIR .'/'. DATADIR .'/'. $hash .'.'. $this->fileExtension;
-		print $destination;
-		if (!move_uploaded_file($this->tmpFile, $destination)) {
-			if (!is_writable(dirname($destination))) {
-				echo dirname($destination) . ' is not writable';
+
+		if (!file_exists($destination)) {
+			if (!move_uploaded_file($this->tmpFile, $destination)) {
+				if (!is_writable(dirname($destination))) {
+					die(dirname($destination) . ' is not writable');
+				} else {
+					die('Upload was not a file.');
+				}
 			} else {
-				die('Upload was not a file.');
+				$db = new database();
+				$db->setupDB();
+				$data = array(
+					'hash' => $hash,
+					'extension' => $this->fileExtension,
+					'originalname' => $this->originalname,
+					'title' => $this->originalname,
+					'description' => $this->originalname
+				);
+				$db->insertImage($data);
 			}
+		} else {
+			die('File already exists');
 		}
 	}
 
-	private function checkFileMIME() {
+	private function checkFileMIME($mime) {
 		$types = array(
 				'image/jpeg',
 				'image/jpg',
 				'image/png'
 			);
 
-		if (in_array(strtolower($this->filemime), $types)) {
+		if (in_array(strtolower($mime), $types)) {
 			// safe file
-			if ($this->filemime == 'image/jpeg' || $this->filemime == 'image/jpg' ) {
+			if ($mime == 'image/jpeg' || $this->fileMIME == 'image/jpg' ) {
 				$this->fileExtension = 'jpg';
-			} else if ($this->filemime == 'image/png') {
+			} else if ($mime == 'image/png') {
 				$this->fileExtension = 'png';
 			}
 			return TRUE;
 		} else {
 			return FALSE;
 		}
-	}
-
-	public function originalFileName() {
-		return $this->filename;
 	}
 }
